move logic from init to views.py Changes to be committed: 	modified:   app/__init__.py 	new file:   app/views.py 	modified:   run.py
