+++ run.py
@@ -1,7 +1,4 @@
-import os
-from app import create_app
-
-app = create_app()
+from app.views import app
 
 if __name__ == '__main__':
     app.run()
